# LibCrypto: Improve GHash / GCM performance