@Hendiadyoin1 2024-08-21
I agree with Ali,
it is currently a bit hard to read and understand whats going on

It generally looks a lot like it's based on a vector algorithm,
Here's a rough draft of what I was able to come up with for the clmul part: <https://godbolt.org/z/G46T8KGr7>