@MarekKnapek 2024-08-22
> I would really, _really_ appreciate some comments on how this blob functions :)
> 
> Nice work!

Added some comments.